RR Update (#16)
* Automatic translation import

Change-Id: Ifcb5e06b65ff0fde5ef6b15c91d028f23cad2c66

* Set APM mode property for RIL to access.

Set APM mode in ADB property so that RIL can access it
to know if the device is in APM in bootup.

Change-Id: Id54667602b1358eaa7a7e45bd77bd0e932569507

* IMS-VT: Conf: Propagate call extras to UI.

Call extras of a conference call are not getting
propagating to UI because the ImsConference class
doesn't listen to the onCallExtras notifcations
of the original connection.

Change-Id: I6f5499e3b18f9de87869cb5c9a0191cd59a70f89
CRs-Fixed: 939132

* SEEK: Change security check for getAtrUsingSubId

- The UID of the third party apk has changed
and is no longer equal to NFC_UID
- Check and throw security exception if apk UID is
not equal to SYSTEM UID and does not have PRIVILEGED
permission

Change-Id: I1c3c339732c149a1f985b44bc9e861442543b96d
CRs-Fixed: 934980

* Add "FDN service is not available" toast message

Display toast message in 'add contact to FDN list' screen,
while airplane mode is enabled.
Add a broadcast receiver for Sim state change and display
the toast if sim is absent.

Change-Id: Ic949ce0563a7458697eb754d4089d7d674cb4a60
CRs-Fixed: 541836

* Mms: Fix SMS is not sent in ECBM unless click send again

When clicking "yes" button on "Emergency Callback Mode" dialog,
EmergencyCallbackModeExitDialog.onDismiss() function will be called
and set the "exit_ecm_result" to false. Thus, after back to
ComposeMessageActivity, the message can't be sent.

Change to listen the AlertDialog's onCancel event instead.

CRs-Fixed: 820011
Change-Id: Icf2e9a4b5e80c7901b0842c13c12a2dded28f50e

* Fix updating FDN using uri without subId issue

This is a side-effect of CR 541836, SIM status changed intent
override static value subId in SubscriptionInfoHelper to -1
because of using wrong key set to get subId from intent,
when inserting FDN it would use uri(content://icc/fdn)
instead of uri(content://icc/fdn/subId/#), thus make telephony
confused to update FDN value to SIM card.

CRs-Fixed: 965758

Change-Id: I07cb6c0cb635c6c9bf3f9bc7c41078d9f65e4e94

* Automatic translation import

* Cleanup for translations

Change-Id: Ie79ffcbf39b601f13d0e87f000560af96a440b7e

* Telephony: Handle PIN locked sim slot swapping between boots.

  The AOSP implementation would always return the default phone
  in an emergency scenario, since we switched to finding the first
  available emergency phone. During a cold swap scenario, at initial
  boot, the modem state -> subscription state may be out of sync while
  it goes through the process of updating the info from the controller.

  Thus, verify the sim state as well as the subscription state prior to
  choosing a phone to emergency dial with.

Change-Id: If28408334d605e71151ce3bc8f4b0c35282d0b37
TICKET: PAELLA-272

* Automatic translation import

Change-Id: I536e1b99a03031289c0031285bdbbe4f0797f2f9

* Automatic translation import

Change-Id: I454d4880bc5bd674f98124881f0c9ef7cb6e1add
Ticket: -

* Automatic translation import

* Automatic translation import

* Automatic translation import

Change-Id: I4665456bdb611436423057479f789c282af28d4b
Ticket: -

* Automatic translation import

Change-Id: I357834ade6a4a5992dee241b1a3d5e2341e62586
Ticket: -

* Telephony: Fix voicemail preference issue

The preferences are all removed and re-created during onResume,
this causes multiple problem, as pick contact list will fire new activity
and when it return back, VoicemailSettingsActivity onResume is called,
making mSubMenuVoicemailSettings pointer to a new object instead of
the one being currently displayed. Therefore saveVoiceMailAndForwardingNumber
is not called. Same situation happens when Phone is temperory put into background
and then back again.

Also during screen orientation change in contact picker, VoicemailSettingsActivity
will be re-created when return, causing mSubMenuVoicemailSettings equal to null and
crash the Phone application

Fix it by avoid un-necessary layout re-creation

CYNGNOS-3184, FEIJ-1504, FEIJ-1505

Change-Id: Iee35b165e5f0aba1aba11c197a513278412b8207

* Automatic translation import

Change-Id: I5b65be7a6c4cc7ab8d52bb01975ef15d13d5d4ed
Ticket: -

* Telephony: reset persisted user nw mode on network factory reset

Ticket: CYNGNOS-3165

Change-Id: If5ff337201daac5156ef7dd044e6929329506fc2
Signed-off-by: Roman Birg <roman@cyngn.com>

* NetworkSelection: Fix double display of radio technology

Some carriers add the technology to their name already so it
is not required to add it again by code.

Change-Id: Ibfa096cf9b62d9054860d9d76d6e5894c68bbf67

* NetworkSelection: Fix for setting shortcut

When adding carrier selection as a setting shortcut, no intent is passed
and the subId can't be extracted  from there.
Use the default subscription Id in that case

Change-Id: Ie70ded06799a1afd3401119e9241fe4f3503e2d6

* NetworkSelection: Cleanup

* Remove unused functions
* Remove redundant "return"

Change-Id: Ie7a121e3f42d4fa557dcd65574b6a0de6b5ddcb2

* Automatic translation import

* EditPhoneNumberPreference: Prevent empty text from being set

These preferences need an entry in the edit text or else
the phone and/or the ui gets put into a weird state where
actions like Call forwarding can appear disabled when they
are not.

Change-Id: I8c70f3fe571651c5cf4cea4696688e391af73911
Ticket: CYNGNOS-3098

* MobileNetworkSettings: Check CarrierConfig on isWorldMode

Add support for checking the CarrierConfig version of the World
Phone config. If this is not present, fall back to the old
deprecated way of looking it up from the Telephony settings.

Not having this prevented the world phone config from working if
you fully migrated to CarrierConfig - you had to have both versions
or you would fail this check

Change-Id: I73939f1055849793d108b02bf034ab79e360c4e4

* Automatic translation import

* Automatic translation import

* Automatic translation import

* Automatic translation import

Change-Id: I0cc7401edc8a7f101a30b786470c8d8d584aacf1
Ticket: -

* Merge tag 'android-6.0.1_r66' into HEAD

Android 6.0.1 release 66

# gpg: Signature made Tue 06 Sep 2016 09:27:12 AM PDT using DSA key ID 9AB10E78
# gpg: Can't check signature: public key not found

* Automatic translation import

* Merge tag 'android-6.0.1_r72' into HEAD

Android 6.0.1 Release 72 (M4B30X)

* Telephony: Fix mobile network settings for CDMA without LTE

Change-Id: I789e30583195011c5f97884a9aa74f66ad90a0cd